Validação para exclusão de usuário.

+++ projeto-pombo/src/main/java/br/senac/projeto_pombo/service/UsuarioService.java
@@ -51,8 +51,10 @@ public class UsuarioService {
 		return repository.save(usuario);
 	}
 
-	public void excluir(Integer id) {
-		
+	public void excluir(Integer id) throws PomboException {
+		if(repository.verificarSePossuiPruu(id)) {
+			throw new PomboException("Não é possível remover usuário que já criou uma postagem.");
+		}
 		repository.deleteById(id);
 	}
